ExponentialMovingAverage now uses float values no matter what data type the initial value is
Commit type: Refactoring

* Using anything other than float32 values doesn't make sense here.

+++ keras_vq_vae.py
@@ -72,8 +72,8 @@ class ExponentialMovingAverage(tfkl.Layer):
 
     def initialize(self, value: tf.Tensor):
         self.initialized.assign(True)
-        self._hidden = tf.Variable(tf.zeros_like(value), trainable=False, name="hidden")
-        self.average = tf.Variable(tf.zeros_like(value), trainable=False, name="average")
+        self._hidden = tf.Variable(tf.zeros_like(value, dtype=tf.float32), trainable=False, name="hidden")
+        self.average = tf.Variable(tf.zeros_like(value, dtype=tf.float32), trainable=False, name="average")
 
 
 class QuantizationLayerEMA(tfkl.Layer):
